+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixes
 - Fix `CheckURLValid` to support `http` and `https`.
 - Fix `SetImage` to update `Image` property.
+- Update `LockDocumentRelayCommand` to use `LockDocumentManager` to support document locking.
 
 ## [0.1.0] / 2025-07-02
 ### Features